Initial: very basic proof-of-concept
Workflow:
  * Frontend requests a shortened url from the API via PUT request
  * API response uses 'content-location' to say what the new url is
  * The shortened URL connects directly to the backend, which uses a 301
    to send them to the right place

CORS is a little hairy at this point, and "*" is not the correct origin
to be allowing.

Note that there's a container dedicated to tests (and one for a test
DB).  This was just an easy way to set things up so that we could run
tests and not potentially interfere with the running app database.