YAY, we get to remove the CORS nonsense
Add a reverse proxy, so that from the perspective of the frontend, there
is just one service running.

I've updated the API service to no longer just run on '/', instead
running on '/redirects', just because this simplfies the nginx
configuration and eliminates any ambiguity about what the frontend is
asking for.